+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [Next]
+
+
 ## [1.5.3] - 2026-01-01
 
 - Venus: Add support for Venus A (VNSA) and Venus D (VNSD) device types