+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## [Next]
+## [1.6.0] - 2026-01-25
 
 - Fix Home Assistant warning when surplus feed-in is unavailable on older HM firmware versions
 - Only encrypt Marstek topic device IDs for HMA, HMF, HMK, and HMJ devices (#231)